tRCP: move server code to the new src/server
