+++ src/apps/personas/ytpersona.router.ts
@@ -3,7 +3,7 @@
 import { TRPCError } from '@trpc/server';
 import { z } from 'zod';
 
-import { createTRPCRouter, publicProcedure } from '~/modules/trpc/trpc.server';
+import { createTRPCRouter, publicProcedure } from '~/server/api/trpc.server';
 
 
 const inputSchema = z.object({
