+++ src/modules/trpc/trpc.client.ts
@@ -9,7 +9,7 @@ import { createTRPCNext } from '@trpc/next';
 import { type inferRouterInputs, type inferRouterOutputs } from '@trpc/server';
 import superjson from 'superjson';
 
-import { type AppRouter } from './trpc.router';
+import type { AppRouter } from '~/server/api/trpc.router';
 
 const getBaseUrl = () => {
   if (typeof window !== 'undefined') return ''; // browser should use relative url
